mail-react: Split the rich-text files into shared and draft-js parts
Prepares to support the TipTap editor in addition to draft-js.
