+++ packages/mail-react/src/blocks/richText/common.ts
@@ -1,19 +1,11 @@
 import type { ReactNode } from "react";
 
 import type { TextStyles, VariantName } from "../../theme/themeTypes.js";
-import type { PropsWithData } from "../helpers/PropsWithData.js";
-
-export interface RichTextBlockData {
-    /** Draft.js raw content state (`{ blocks, entityMap }`), as produced by the CMS RichText block. */
-    draftContent: unknown;
-}
-
-export type RichTextBlockProps = PropsWithData<RichTextBlockData>;
 
 export type RichTextListKind = "unordered" | "ordered";
 
 /**
- * Styling and list kind for all draft blocks of one type, on top of the base theme text styles.
+ * Styling and list kind for all text blocks of one type, on top of the base theme text styles.
  *
  * Style props accept plain values only. For responsive styling, use a theme
  * variant, or set a `className` and register responsive CSS via `registerStyles`.
@@ -26,7 +18,7 @@ export type RichTextBlockTypeProps = Omit<TextStyles, "bottomSpacing"> & {
      */
     variant?: VariantName;
     className?: string;
-    /** Renders every draft block of this type as a list of this kind. */
+    /** Renders every text block of this type as a list of this kind. */
     list?: RichTextListKind;
 };
 
@@ -38,37 +30,8 @@ export type RichTextBlockTypeProps = Omit<TextStyles, "bottomSpacing"> & {
 export type RichTextLinkHrefResolver<TProps = unknown> = (props: TProps) => string | undefined;
 
 /**
- * Renders the text spanned by one draft-js inline style (e.g. `BOLD`).
+ * Renders the text spanned by one piece of inline formatting.
  *
  * `key` must be set on the returned element's root.
  */
 export type RichTextInlineRenderer = (children: ReactNode, options: { key: string }) => ReactNode;
-
-export interface CreateRichTextBlockOptions<TLinkTypes extends Record<string, unknown> = Record<string, unknown>> {
-    /**
-     * Maps draft block types (e.g. `"header-one"`, `"paragraph-standard"`) to how
-     * they render.
-     *
-     * Unmapped block types render as paragraphs with the base theme text styles,
-     * apart from `unordered-list-item` and `ordered-list-item`.
-     */
-    blockTypes?: Record<string, RichTextBlockTypeProps>;
-    /**
-     * Maps the application's link block types within `LINK` entities to a
-     * resolver returning the link's href.
-     *
-     * Merged on top of the built-in `external` link type. Link types without
-     * a resolver render their text without a link.
-     */
-    linkTypes?: { [TLinkType in keyof TLinkTypes]: RichTextLinkHrefResolver<TLinkTypes[TLinkType]> };
-    /**
-     * Maps draft-js inline style names to renderers, keyed by the style name as
-     * it appears in the content's `inlineStyleRanges`.
-     *
-     * Merged on top of the built-in styles (`BOLD`, `ITALIC`, `SUB`, `SUP`,
-     * `STRIKETHROUGH`): use it to override a built-in style, or to render a
-     * custom inline style the application defines in its RTE (e.g. `HIGHLIGHT`),
-     * which has no built-in renderer.
-     */
-    inline?: Record<string, RichTextInlineRenderer>;
-}
