+++ packages/mail-react/src/blocks/richText/draftJs/__stories__/MjmlRichTextBlock.stories.tsx
@@ -1,12 +1,12 @@
 import { MjmlColumn } from "@faire/mjml-react";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
-import { MjmlSection } from "../../../components/section/MjmlSection.js";
-import { registerStyles } from "../../../styles/registerStyles.js";
-import { createTheme } from "../../../theme/createTheme.js";
-import { ThemeProvider } from "../../../theme/ThemeProvider.js";
-import type { Theme } from "../../../theme/themeTypes.js";
-import { css } from "../../../utils/css.js";
+import { MjmlSection } from "../../../../components/section/MjmlSection.js";
+import { registerStyles } from "../../../../styles/registerStyles.js";
+import { createTheme } from "../../../../theme/createTheme.js";
+import { ThemeProvider } from "../../../../theme/ThemeProvider.js";
+import type { Theme } from "../../../../theme/themeTypes.js";
+import { css } from "../../../../utils/css.js";
 import { createRichTextBlock } from "../createRichTextBlock.js";
 import {
     bulletedListBlockData,
