+++ packages/mail-react/src/blocks/richText/draftJs/__stories__/exampleBlockData.ts
