+++ packages/mail-react/src/index.ts
@@ -8,8 +8,8 @@ export type { HtmlPixelImageBlockProps } from "./blocks/pixelImage/HtmlPixelImag
 export { HtmlPixelImageBlock } from "./blocks/pixelImage/HtmlPixelImageBlock.js";
 export type { MjmlPixelImageBlockProps } from "./blocks/pixelImage/MjmlPixelImageBlock.js";
 export { MjmlPixelImageBlock } from "./blocks/pixelImage/MjmlPixelImageBlock.js";
-export type { CreateRichTextBlockOptions, RichTextBlockProps } from "./blocks/richText/common.js";
-export { createRichTextBlock } from "./blocks/richText/createRichTextBlock.js";
+export type { CreateRichTextBlockOptions, RichTextBlockProps } from "./blocks/richText/draftJs/common.js";
+export { createRichTextBlock } from "./blocks/richText/draftJs/createRichTextBlock.js";
 export type { HtmlButtonProps } from "./components/button/HtmlButton.js";
 export { HtmlButton } from "./components/button/HtmlButton.js";
 export type { MjmlButtonProps } from "./components/button/MjmlButton.js";
